fix(video-editor): security hardening, code cleanup, and API smoke tests
Security:
- Sanitize upload filenames to prevent path traversal
- Validate script paths against project directory tree
- Support media unassignment (empty string) for undo sync

Code cleanup:
- Prune completed download tasks (cap at 20) to prevent unbounded growth
- Deduplicate narration: REST endpoint now delegates to service layer like WS
- Scope narration task state to project dir to avoid stale cross-project data
- Forward tts_engine/tts_voice params in POST /produce endpoint
- Remove dead PreviewPanel component
- Remove deprecated selectedSegmentId shim from store
- Fix duration_seconds schema from int to float
- Make undo/redo API-call-first so failures don't corrupt history stacks
- Type all API client functions (StatusResponse, NarrationStatus)
- Add try/catch to WebSocket JSON.parse

Tests:
- Add test_api.py with 32 FastAPI TestClient smoke tests covering all route
  groups, security boundaries, and validation

+++ bee-content/video-editor/src/bee_video_editor/api/schemas.py
@@ -26,7 +26,7 @@ class SegmentSchema(BaseModel):
     section: str
     section_time: str
     subsection: str
-    duration_seconds: int
+    duration_seconds: float
     visual: list[LayerEntrySchema]
     audio: list[LayerEntrySchema]
     overlay: list[LayerEntrySchema]
